move `add_root_for_windows` from root of `util` crate to `util::paths`
