+++ crates/util/src/util.rs
@@ -760,16 +760,6 @@ pub fn word_consists_of_emojis(s: &str) -> bool {
     prev_end == s.len()
 }
 
-#[cfg(any(test, feature = "test-support"))]
-pub fn add_root_for_windows(path: &str) -> String {
-    if cfg!(windows) {
-        assert!(path.starts_with("/"));
-        format!("C:{}", path)
-    } else {
-        path.to_string()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
